requirements.txt
django~=4.0.4
djangorestframework~=3.13.1
django-filter~=21.1
Serializer~=0.2.1